requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
pycryptodome==3.21.0
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
openai>=1.12.0
python-environ>=0.4.54
requests>=2.31.0